requirements.txt
Flask==2.2.2
Werkzeug==2.2.3
pandas==1.5.3
joblib==1.5.0
requests==2.32.3
python-decouple==3.8
flask-cors==5.0.1
opencv-python==4.11.0.86
numpy==1.23.5
scikit-learn==1.2.2
mysql-connector-python==9.3.0
flask-session==0.8.0
bcrypt==4.3.0
flask-mysqldb==2.0.0
PyJWT==2.10.1
python-dotenv==1.1.0
click==8.1.8
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysqlclient==2.2.7
scipy==1.15.2
threadpoolctl==3.6.0
deepface==0.0.79
tensorflow==2.12.0

# Optional - only needed if you want to use facial emotion detection
# deepface
# tensorflow
